+++ govtool/frontend/src/components/organisms/GovernanceActionDetailsCardData.tsx
@@ -18,6 +18,7 @@ import {
   getProposalTypeLabel,
   filterUpdatableProtocolParams,
   filterOutNullParams,
+  getFullGovActionId,
 } from "@utils";
 import { MetadataValidationStatus, ProposalData } from "@models";
 import { GovernanceActionType } from "@/types/governanceAction";
@@ -125,10 +126,10 @@ export const GovernanceActionDetailsCardData = ({
   };
 
   const label = getProposalTypeLabel(type);
-  const govActionId = `${index}${txHash}`;
+  const govActionId = getFullGovActionId(txHash, index);
   const prevGovActionId =
     prevGovActionIndex && prevGovActionTxHash
-      ? `${prevGovActionIndex}${prevGovActionTxHash}`
+      ? getFullGovActionId(prevGovActionTxHash, prevGovActionIndex)
       : null;
 
   const tabs = useMemo(